moved delete route to source_app/views.py
